feat: decouple image builder from worker
