+++ backend/builder/tasks/__init__.py
@@ -1,0 +1,3 @@
+from builder.tasks.tasks_build_image import build_image
+
+__all__ = ["build_image"]
